requirements.txt
torch~=2.3.1+cu121
PyQt6~=6.7.1
opencv-python~=4.10.0.84
numpy~=1.26.4
pillow~=10.4.0
torchvision~=0.18.1+cu121
transformers~=4.42.4
scikit-image~=0.24.0
PyGetWindow~=0.0.9
mss~=9.0.1
requests~=2.32.3